+++ CHANGELOG.md
@@ -8,6 +8,7 @@ how this file is formatted and how the process works.
 ### Fixed
 - Avantio: add handling of `check_in_too_near` response code
 - Avantio: mark first MARGIN = 3 days as unavailable
+- Avantio: add House(Casa) property type
 
 ### Added
 - Ciirus: add new spec for cancel action when `CancelBookingResponse` contains API error